+++ src/ext/java/org/opentripplanner/ext/readiness_endpoint/ActuatorAPI.java
@@ -1,5 +1,7 @@
 package org.opentripplanner.ext.readiness_endpoint;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.opentripplanner.standalone.server.OTPServer;
 import org.opentripplanner.standalone.server.Router;
 
@@ -7,11 +9,14 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.Collection;
 
 
 @Path("/actuators")
 @Produces(MediaType.APPLICATION_JSON) // One @Produces annotation for all endpoints.
 public class ActuatorAPI {
+
+    private static final Logger LOG = LoggerFactory.getLogger(ActuatorAPI.class);
 
     private final Router router;
 
@@ -44,6 +49,19 @@ public class ActuatorAPI {
     @GET
     @Path("/health")
     public Response health() {
+        if (router.graph.updaterManager != null) {
+            Collection<String> waitingUpdaters = router.graph.updaterManager.waitingUpdaters();
+
+            if (!waitingUpdaters.isEmpty()) {
+                LOG.info("Graph ready, waiting for updaters: {}", waitingUpdaters);
+                throw new WebApplicationException(Response
+                    .status(Response.Status.NOT_FOUND)
+                    .entity("Graph ready, waiting for updaters: " + waitingUpdaters + "\n")
+                    .type("text/plain")
+                    .build());
+            }
+        }
+
         return Response.status(Response.Status.OK).entity(
             "{\n"
             + "  \"status\" : \"UP\""