+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/mappers/TransfersMapper.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.routing.algorithm.raptor.transit.mappers;
 
 import org.opentripplanner.model.Stop;
+import org.opentripplanner.routing.algorithm.raptor.transit.StopIndexForRaptor;
 import org.opentripplanner.routing.algorithm.raptor.transit.Transfer;
 import org.opentripplanner.routing.edgetype.SimpleTransfer;
 import org.opentripplanner.routing.graph.Edge;
@@ -22,8 +23,8 @@ class TransfersMapper {
 
         List<List<Transfer>> transferByStopIndex = new ArrayList<>();
 
-        for (int i = 0; i < stopIndex.stops.size(); ++i) {
-            Stop stop = stopIndex.stops.get(i);
+        for (int i = 0; i < stopIndex.stopsByIndex.size(); ++i) {
+            Stop stop = stopIndex.stopsByIndex.get(i);
             ArrayList<Transfer> list = new ArrayList<>();
             transferByStopIndex.add(list);