+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/request/RaptorRoutingRequestTransitDataCreator.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.routing.algorithm.raptor.transit.request;
 
+import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.routing.algorithm.raptor.transit.TransitLayer;
 import org.opentripplanner.routing.algorithm.raptor.transit.TripPattern;
 import org.opentripplanner.routing.algorithm.raptor.transit.TripPatternForDate;
@@ -50,7 +51,7 @@ class RaptorRoutingRequestTransitDataCreator {
       int dayRange, TraverseModeSet transitModes
   ) {
 
-    List<Map<Integer, TripPatternForDate>> tripPatternForDates = getTripPatternsForDateRange(
+    List<Map<FeedScopedId, TripPatternForDate>> tripPatternForDates = getTripPatternsForDateRange(
         dayRange,
         transitModes
     );
@@ -60,10 +61,10 @@ class RaptorRoutingRequestTransitDataCreator {
     return createTripPatternsPerStop(tripPatternForDateList, transitLayer.getStopCount());
   }
 
-  private List<Map<Integer, TripPatternForDate>> getTripPatternsForDateRange(
+  private List<Map<FeedScopedId, TripPatternForDate>> getTripPatternsForDateRange(
       int dayRange, TraverseModeSet transitModes
   ) {
-    List<Map<Integer, TripPatternForDate>> tripPatternForDates = new ArrayList<>();
+    List<Map<FeedScopedId, TripPatternForDate>> tripPatternForDates = new ArrayList<>();
 
     // Start at yesterdays date to account for trips that cross midnight. This is also
     // accounted for in TripPatternForDates.
@@ -87,12 +88,12 @@ class RaptorRoutingRequestTransitDataCreator {
    * performance for searching, as each TripPattern is searched only once per round.
    */
   static List<TripPatternForDates> merge(
-      ZonedDateTime searchStartTime, List<Map<Integer, TripPatternForDate>> tripPatternForDateList
+      ZonedDateTime searchStartTime, List<Map<FeedScopedId, TripPatternForDate>> tripPatternForDateList
   ) {
     List<TripPatternForDates> combinedList = new ArrayList<>();
 
     // Extract all distinct TripPatterns across the search days
-    Map<Integer, TripPattern> allTripPatternsById = tripPatternForDateList
+    Map<FeedScopedId, TripPattern> allTripPatternsById = tripPatternForDateList
         .stream()
         .flatMap(t -> t.values().stream())
         .map(TripPatternForDate::getTripPattern)
@@ -100,11 +101,11 @@ class RaptorRoutingRequestTransitDataCreator {
         .collect(Collectors.toMap(TripPattern::getId, t -> t));
 
     // For each TripPattern, time expand each TripPatternForDate and merge into a single TripPatternForDates
-    for (Map.Entry<Integer, TripPattern> patternEntry : allTripPatternsById.entrySet()) {
+    for (Map.Entry<FeedScopedId, TripPattern> patternEntry : allTripPatternsById.entrySet()) {
       List<TripPatternForDate> tripPatterns = new ArrayList<>();
       List<Integer> offsets = new ArrayList<>();
 
-      for (Map<Integer, TripPatternForDate> tripPatternById : tripPatternForDateList) {
+      for (Map<FeedScopedId, TripPatternForDate> tripPatternById : tripPatternForDateList) {
         TripPatternForDate tripPatternForDate = tripPatternById.get(patternEntry.getKey());
         if (tripPatternForDate != null) {
           tripPatterns.add(tripPatternForDate);
@@ -134,7 +135,7 @@ class RaptorRoutingRequestTransitDataCreator {
     return result;
   }
 
-  private static Map<Integer, TripPatternForDate> listActiveTripPatterns(
+  private static Map<FeedScopedId, TripPatternForDate> listActiveTripPatterns(
       TransitLayer transitLayer, LocalDate date, TraverseModeSet transitModes
   ) {