+++ src/main/java/org/opentripplanner/updater/stoptime/WebsocketGtfsRealtimeUpdater.java
@@ -181,4 +181,9 @@ public class WebsocketGtfsRealtimeUpdater implements GraphUpdater {
             }
         }
     }
+
+    @Override
+    public String getName() {
+        return "WebsocketGtfsRealtimeUpdater";
+    }
 }