+++ src/test/java/org/opentripplanner/routing/algorithm/raptor/transit/mappers/StopIndexForRaptorTest.java
@@ -3,6 +3,7 @@ package org.opentripplanner.routing.algorithm.raptor.transit.mappers;
 import org.junit.Test;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Stop;
+import org.opentripplanner.routing.algorithm.raptor.transit.StopIndexForRaptor;
 
 import java.util.Arrays;
 import java.util.List;